Refactoring to prepare for VerifyEmailSignUpActivity
