+++ app/src/main/java/com/ncsuadc/marigold_android/ui/home/shared/composables.kt
@@ -1,0 +1,77 @@
+package com.ncsuadc.marigold_android.ui.home.shared
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+// Should this be in the theme or typography file?
+val SIGN_UP_TITLE_STYLE = TextStyle(fontWeight = FontWeight.ExtraBold, fontSize = 35.sp)
+
+@Composable
+fun TitleText(modifier: Modifier = Modifier) {
+    Text(
+        buildAnnotatedString {
+            withStyle(
+                SpanStyle(
+                    brush = Brush.linearGradient(
+                        // If we plan to use these colors more often we should make them a constant
+                        // somewhere/add it to the theme. I'm not sure how to do this, however.
+                        colors = listOf(Color(0xffffe501), Color(0xffffb320))
+                    )
+                )
+            ) {
+                append("Marigold ")
+            }
+            append("@\n")
+            withStyle(SpanStyle(color = Color(0xffff3737))) {
+                append("NC State University")
+            }
+        },
+        style = SIGN_UP_TITLE_STYLE,
+        modifier = modifier
+    )
+}
+
+// I took some liberties with the gradient button, but I think it looks better this way.
+// Feel free to change it if you disagree.
+@Composable
+fun GradientButton(
+    gradient : Brush,
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit = { },
+    content: @Composable () -> Unit
+) {
+    Button(
+        modifier = modifier,
+        colors = ButtonDefaults.buttonColors(containerColor = Color.Transparent),
+        shape = MaterialTheme.shapes.medium,
+        contentPadding = PaddingValues(),
+        onClick = { onClick() },
+    ) {
+        Row(
+            modifier = Modifier
+                .background(gradient)
+                .padding(horizontal = 20.dp, vertical = 15.dp),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            content()
+        }
+    }
+}
