+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,7 @@
     <string name="title_activity_home">HomeActivity</string>
     <string name="title_activity_test">TestActivity</string>
     <string name="title_activity_sign_up">SignUpActivity</string>
+    <string name="title_activity_verify_email_sign_up">VerifyEmailSignUpActivity</string>
+<!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
